+++ backend/routes/schedulerRouter.js
@@ -1,12 +1,12 @@
-import express from 'express';
-import cron from './scheduler.js';
+// import express from 'express';
+// import cron from './scheduler.js';
 
 
-const router = express.Router();
+// const router = express.Router();
 
-router.get('/start-cron', (req, res) => {
-  cron.start();
-  res.send('Cron job started manually.');
-});
+// router.get('/start-cron', (req, res) => {
+//   cron.start();
+//   res.send('Cron job started manually.');
+// });
 
-export default router;
+// export default router;
